Move Scene to engine file
